+++ android/src/main/java/com/synerise/sdk/react/RNContent.java
@@ -13,8 +13,10 @@ import com.synerise.sdk.content.model.Audience;
 import com.synerise.sdk.content.model.DocumentsApiQuery;
 import com.synerise.sdk.content.model.DocumentsApiQueryType;
 import com.synerise.sdk.content.model.ScreenViewResponse;
+import com.synerise.sdk.content.model.document.Document;
 import com.synerise.sdk.content.model.recommendation.RecommendationRequestBody;
 import com.synerise.sdk.content.model.recommendation.RecommendationResponse;
+import com.synerise.sdk.content.model.screenview.ScreenView;
 import com.synerise.sdk.content.widgets.dataModel.Recommendation;
 import com.synerise.sdk.core.listeners.DataActionListener;
 import com.synerise.sdk.core.net.IDataApiCall;
@@ -34,11 +36,13 @@ import javax.annotation.Nonnull;
 
 public class RNContent extends RNBaseModule {
 
-    IDataApiCall<Object> getDocumentApiCall;
-    IDataApiCall<List<Object>> getDocumentsApiCall;
+    private IDataApiCall<Object> getDocumentApiCall;
+    private IDataApiCall<List<Object>> getDocumentsApiCall;
+    private IDataApiCall<Document> generateDocumentApiCall;
     private final String ISO8601_FORMAT = "yyyy-MM-dd'T'kk:mm:ss.SSS'Z'";
     private IDataApiCall<RecommendationResponse> getRecommendationsApiCall;
     private IDataApiCall<ScreenViewResponse> getScreenViewApiCall;
+    private IDataApiCall<ScreenView> generateScreenViewApiCall;
     private Gson gson = new Gson();
 
     public RNContent(ReactApplicationContext reactApplicationContext) {
@@ -185,6 +189,111 @@ public class RNContent extends RNBaseModule {
         });
     }
 
+    @ReactMethod
+    public void getRecommendationsV2(ReadableMap recommendationsOptions, Callback callback) {
+        String productId = recommendationsOptions.hasKey("productID") ? recommendationsOptions.getString("productID") : "";
+        String slugName = recommendationsOptions.hasKey("slug") ? recommendationsOptions.getString("slug") : "";
+
+        if (getRecommendationsApiCall != null) getRecommendationsApiCall.cancel();
+
+        RecommendationRequestBody recommendationRequestBody = new RecommendationRequestBody();
+        recommendationRequestBody.setProductId(productId);
+
+        getRecommendationsApiCall = Content.getRecommendationsV2(slugName, recommendationRequestBody);
+        getRecommendationsApiCall.execute(new DataActionListener<RecommendationResponse>() {
+            @Override
+            public void onDataAction(RecommendationResponse responseBody) {
+                WritableMap recommendationMap = Arguments.createMap();
+
+                recommendationMap.putString("campaignHash", responseBody.getCampaignHash());
+                recommendationMap.putString("campaignId", responseBody.getCampaignId());
+                recommendationMap.putString("schema", responseBody.getSchema());
+                recommendationMap.putString("slug", responseBody.getSlug());
+                recommendationMap.putString("uuid", responseBody.getUuid());
+
+                recommendationMap.putArray("items", recommendationToWritableArray(responseBody.getRecommendations()));
+                executeSuccessCallbackResponse(callback, recommendationMap, null);
+            }
+        }, new DataActionListener<ApiError>() {
+            @Override
+            public void onDataAction(ApiError apiError) {
+                executeFailureCallbackResponse(callback, null, apiError);
+            }
+        });
+    }
+
+    @ReactMethod
+    public void generateDocument(String slug, Callback callback) {
+        if (generateDocumentApiCall != null) generateDocumentApiCall.cancel();
+
+        generateDocumentApiCall = Content.generateDocument(slug);
+        generateDocumentApiCall.execute(new DataActionListener<Document>() {
+
+            @Override
+            public void onDataAction(Document document) {
+                WritableMap documentMap = Arguments.createMap();
+                documentMap.putString("uuid", document.getUuid());
+                documentMap.putString("slug", document.getSlug());
+                documentMap.putString("schema", document.getSchema());
+                try {
+                    String jsonObject = gson.toJson(document);
+                    WritableMap objectMap = convertJsonToMap(new JSONObject(jsonObject));
+                    documentMap.putMap("content", objectMap);
+                    executeSuccessCallbackResponse(callback, documentMap, null);
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            }
+        }, new DataActionListener<ApiError>() {
+            @Override
+            public void onDataAction(ApiError apiError) {
+                executeFailureCallbackResponse(callback, null, apiError);
+            }
+        });
+    }
+
+    @ReactMethod
+    public void generateScreenView(String feedSlug, Callback callback) {
+        if (generateScreenViewApiCall != null) generateScreenViewApiCall.cancel();
+        generateScreenViewApiCall = Content.generateScreenView(feedSlug);
+        generateScreenViewApiCall.execute(
+                new DataActionListener<ScreenView>() {
+                    @Override
+                    public void onDataAction(ScreenView data) {
+                        WritableMap screenViewMap = Arguments.createMap();
+                        screenViewMap.putMap("audience", audienceNewScreenViewsToWritableMap(data.getAudience()));
+                        screenViewMap.putString("identifier", data.getId());
+                        screenViewMap.putString("hash", data.getHash());
+                        screenViewMap.putString("path", data.getPath());
+                        screenViewMap.putString("name", data.getName());
+                        screenViewMap.putInt("priority", data.getPriority());
+                        screenViewMap.putMap("data", screenViewDataToWritableMap(data.getData()));
+
+                        try {
+                            Date createdAtDate = new SimpleDateFormat(ISO8601_FORMAT, Locale.getDefault()).parse(data.getCreatedAt());
+                            Date updatedAtDate = new SimpleDateFormat(ISO8601_FORMAT, Locale.getDefault()).parse(data.getUpdatedAt());
+
+                            if (createdAtDate != null) {
+                                screenViewMap.putDouble("createdAt", createdAtDate.getTime());
+                            }
+                            if (updatedAtDate != null) {
+                                screenViewMap.putDouble("updatedAt", updatedAtDate.getTime());
+                            }
+                        } catch (ParseException e) {
+                            e.printStackTrace();
+                        }
+
+                        executeSuccessCallbackResponse(callback, screenViewMap, null);
+                    }
+                }, new DataActionListener<ApiError>() {
+                    @Override
+                    public void onDataAction(ApiError apiError) {
+                        executeFailureCallbackResponse(callback, null, apiError);
+                    }
+                }
+        );
+    }
+
     private WritableArray recommendationToWritableArray(List<Recommendation> array) {
         WritableArray writableArray = Arguments.createArray();
 
@@ -223,6 +332,16 @@ public class RNContent extends RNBaseModule {
         List<String> idsList = audience.getIds();
         audienceMap.putString("query", audience.getQuery());
         audienceMap.putArray("IDs", idsList != null ? listOfStringsIntoWritableArray(idsList) : null);
+
+        return audienceMap;
+    }
+
+    private WritableMap audienceNewScreenViewsToWritableMap(com.synerise.sdk.content.model.screenview.Audience audience) {
+        WritableMap audienceMap = Arguments.createMap();
+        List<String> segments = audience.getSegments();
+        audienceMap.putArray("segments", segments != null ? listOfStringsIntoWritableArray(segments) : null);
+        audienceMap.putString("targetType", audience.getTargetType());
+        audienceMap.putString("query", audience.getQuery());
 
         return audienceMap;
     }