+++ ios/ReactNativeSynerise/Modules/RNContent.m
@@ -66,6 +66,21 @@ RCT_EXPORT_MODULE();
 
 #pragma mark - JS Mapping
 
+- (nullable NSDictionary *)dictionaryWithDocument:(SNRDocument *)model {
+    if (model != nil) {
+        NSMutableDictionary *dictionary = [@{} mutableCopy];
+        
+        [dictionary setString:model.identifier forKey:@"uuid"];
+        [dictionary setString:model.slug forKey:@"slug"];
+        [dictionary setString:model.schema forKey:@"schema"];
+        [dictionary setDictionary:model.content forKey:@"content"];
+        
+        return dictionary;
+    }
+    
+    return nil;
+}
+
 - (nullable NSDictionary *)dictionaryWithRecommendationResponse:(SNRRecommendationResponse *)model {
     if (model != nil) {
         NSMutableDictionary *dictionary = [@{} mutableCopy];
@@ -128,12 +143,48 @@ RCT_EXPORT_MODULE();
     return nil;
 }
 
+- (nullable NSDictionary *)dictionaryWithScreenView:(SNRScreenView *)model {
+    if (model != nil) {
+        NSMutableDictionary *dictionary = [@{} mutableCopy];
+        
+        [dictionary setString:model.identifier forKey:@"identifier"];
+        [dictionary setString:model.name forKey:@"name"];
+        [dictionary setString:model.hashString forKey:@"hashString"];
+        [dictionary setString:model.path forKey:@"path"];
+        [dictionary setInteger:model.priority forKey:@"priority"];
+        [dictionary setDictionary:[self dictionaryWithScreenViewAudienceInfo:model.audience] forKey:@"audience"];
+        
+        [dictionary setGenericObject:model.data forKey:@"data"];
+        
+        [dictionary setDate:model.createdAt forKey:@"createdAt"];
+        [dictionary setDate:model.updatedAt forKey:@"updatedAt"];
+        
+        return dictionary;
+    }
+    
+    return nil;
+}
+
 - (nullable NSDictionary *)dictionaryWithScreenViewAudience:(SNRScreenViewAudience *)model {
     if (model != nil) {
         NSMutableDictionary *dictionary = [@{} mutableCopy];
         
         [dictionary setArray:model.IDs forKey:@"IDs"];
         [dictionary setString:model.query forKey:@"query"];
+        
+        return dictionary;
+    }
+    
+    return nil;
+}
+
+- (nullable NSDictionary *)dictionaryWithScreenViewAudienceInfo:(SNRScreenViewAudienceInfo *)model {
+    if (model != nil) {
+        NSMutableDictionary *dictionary = [@{} mutableCopy];
+        
+        [dictionary setArray:model.segments forKey:@"segments"];
+        [dictionary setString:model.query forKey:@"query"];
+        [dictionary setString:model.targetType forKey:@"targetType"];
         
         return dictionary;
     }
@@ -150,6 +201,22 @@ RCT_EXPORT_METHOD(getDocument:(NSString *)slug response:(RCTResponseSenderBlock)
     [SNRContent getDocument:slug success:^(NSDictionary *document) {
         if (document != nil) {
             [self executeSuccessCallbackResponse:response data:document];
+        } else {
+            [self executeDefaultFailureCallbackResponse:response];
+        }
+    } failure:^(NSError *error) {
+        [self executeFailureCallbackResponse:response error:error];
+    }];
+}
+
+//generateDocument(slug: string, onSuccess: (document: Document) => void, onError: (error: Error) => void)
+
+RCT_EXPORT_METHOD(generateDocument:(NSString *)slug response:(RCTResponseSenderBlock)response)
+{
+    [SNRContent generateDocument:slug success:^(SNRDocument *document) {
+        NSDictionary *documentDictionary = [self dictionaryWithDocument:document];
+        if (documentDictionary != nil) {
+            [self executeSuccessCallbackResponse:response data:documentDictionary];
         } else {
             [self executeDefaultFailureCallbackResponse:response];
         }
@@ -176,13 +243,36 @@ RCT_EXPORT_METHOD(getDocuments:(NSDictionary *)dictionary response:(RCTResponseS
     }
 }
 
-//getRecommendations(recommendationOptions: RecommendationOptions, onSuccess: (recommendationResponse: RecommendationResponse) => void, onError: (error: Error) => void)
+//getRecommendations(options: RecommendationOptions, onSuccess: (recommendationResponse: RecommendationResponse) => void, onError: (error: Error) => void)
 
 RCT_EXPORT_METHOD(getRecommendations:(NSDictionary *)dictionary response:(RCTResponseSenderBlock)response)
 {
     SNRRecommendationOptions *recommendationOptions = [self modelRecommendationOptionsWithDictionary:dictionary];
     if (recommendationOptions != nil) {
         [SNRContent getRecommendations:recommendationOptions success:^(SNRRecommendationResponse *recommendationResponse) {
+            if (recommendationResponse != nil) {
+                NSDictionary *recommendationResponseDictionary = [self dictionaryWithRecommendationResponse:recommendationResponse];
+                if (recommendationResponseDictionary != nil) {
+                    [self executeSuccessCallbackResponse:response data:recommendationResponseDictionary];
+                } else {
+                    [self executeDefaultFailureCallbackResponse:response];
+                }
+            } else {
+                [self executeDefaultFailureCallbackResponse:response];
+            }
+        } failure:^(NSError *error) {
+            [self executeFailureCallbackResponse:response error:error];
+        }];
+    }
+}
+
+//getRecommendationsV2(options: RecommendationOptions, onSuccess: (recommendationResponse: RecommendationResponse)
+
+RCT_EXPORT_METHOD(getRecommendationsV2:(NSDictionary *)dictionary response:(RCTResponseSenderBlock)response)
+{
+    SNRRecommendationOptions *recommendationOptions = [self modelRecommendationOptionsWithDictionary:dictionary];
+    if (recommendationOptions != nil) {
+        [SNRContent getRecommendationsV2:recommendationOptions success:^(SNRRecommendationResponse *recommendationResponse) {
             if (recommendationResponse != nil) {
                 NSDictionary *recommendationResponseDictionary = [self dictionaryWithRecommendationResponse:recommendationResponse];
                 if (recommendationResponseDictionary != nil) {
@@ -208,6 +298,24 @@ RCT_REMAP_METHOD(getScreenView, getScreenViewWithResponse:(RCTResponseSenderBloc
             NSDictionary *screenViewResponseDictionary = [self dictionaryWithScreenViewResponse:screenViewResponse];
             if (screenViewResponseDictionary != nil) {
                 [self executeSuccessCallbackResponse:response data:screenViewResponseDictionary];
+            } else {
+                [self executeDefaultFailureCallbackResponse:response];
+            }
+        }
+    } failure:^(NSError *error) {
+        [self executeFailureCallbackResponse:response error:error];
+    }];
+}
+
+//generateScreenView(feedSlug: string, onSuccess: (screenView: ScreenView) => void, onError: (error: Error) => void)
+
+RCT_EXPORT_METHOD(generateScreenView:(NSString *)feedSlug response:(RCTResponseSenderBlock)response)
+{
+    [SNRContent generateScreenView:feedSlug success:^(SNRScreenView *screenView) {
+        if (screenView != nil) {
+            NSDictionary *screenViewDictionary = [self dictionaryWithScreenView:screenView];
+            if (screenViewDictionary != nil) {
+                [self executeSuccessCallbackResponse:response data:screenViewDictionary];
             } else {
                 [self executeDefaultFailureCallbackResponse:response];
             }