+++ lib/classes/models/Content/Document.d.ts
@@ -1,0 +1,15 @@
+import { BaseModel } from '../BaseModel';
+interface IDocument {
+    uuid: string;
+    slug: string;
+    schema: string;
+    content: object;
+}
+declare class Document extends BaseModel {
+    uuid: string;
+    slug: string;
+    schema: string;
+    content: object;
+    constructor(modelObject: IDocument);
+}
+export { IDocument, Document };