+++ lib/classes/models/Content/ScreenViewAudienceInfo.d.ts
@@ -1,0 +1,13 @@
+import { BaseModel } from '../BaseModel';
+interface IScreenViewAudienceInfo {
+    segments?: Array<string>;
+    query?: string;
+    targetType?: string;
+}
+declare class ScreenViewAudienceInfo extends BaseModel {
+    segments?: Array<string>;
+    query?: string;
+    targetType?: string;
+    constructor(modelObject: IScreenViewAudienceInfo);
+}
+export { IScreenViewAudienceInfo, ScreenViewAudienceInfo };