+++ lib/main/modules/ContentModule.d.ts
@@ -4,6 +4,8 @@ import { RecommendationOptions } from '../../classes/models/Content/Recommendati
 import { RecommendationResponse } from '../../classes/models/Content/RecommendationResponse';
 import { Error } from '../..';
 import { ScreenViewResponse } from '../../config/import_models';
+import { ScreenView } from '../../classes/models/Content/ScreenView';
+import { Document } from '../../classes/models/Content/Document';
 declare class ContentModule extends Module {
     private static _instance;
     static instance(): ContentModule;
@@ -15,8 +17,18 @@ declare class ContentModule extends Module {
      * @param onSuccess Function to be executed when the operation finishes successfully
      * @param onError Function to be executed when the operation finishes unsuccessfully
      *
+     * @deprecated Use the new `Content.generateDocument(slug:onSuccess:onError)` method instead.
      */
     getDocument(slug: string, onSuccess: (document: object) => void, onError: (error: Error) => void): void;
+    /**
+     * This method generates the document that is defined for the provided slug.
+     *
+     * @param slug Identifies a specific document
+     * @param onSuccess Function to be executed when the operation finishes successfully
+     * @param onError Function to be executed when the operation finishes unsuccessfully
+     *
+     */
+    generateDocument(slug: string, onSuccess: (document: Document) => void, onError: (error: Error) => void): void;
     /**
      * This method gets documents that are defined for parameters provided in the query object.
      *
@@ -24,6 +36,7 @@ declare class ContentModule extends Module {
      * @param onSuccess Function to be executed when the operation finishes successfully
      * @param onError Function to be executed when the operation finishes unsuccessfully
      *
+     * @deprecated This method is deprecated.
      */
     getDocuments(apiQuery: DocumentsApiQuery, onSuccess: (documents: Array<object>) => void, onError: (error: Error) => void): void;
     /**
@@ -33,15 +46,33 @@ declare class ContentModule extends Module {
      * @param onSuccess Function to be executed when the operation finishes successfully
      * @param onError Function to be executed when the operation finishes unsuccessfully
      *
+     * @deprecated Use the new `Content.getRecommendationsV2(options:onSuccess:onError)` method instead.
      */
     getRecommendations(options: RecommendationOptions, onSuccess: (recommendationResponse: RecommendationResponse) => void, onError: (error: Error) => void): void;
+    /**
+     * This method gets recommendations that are defined for the options provided.
+     *
+     * @param options `RecommendationOptions` object providing parameters for recommendations
+     * @param onSuccess Function to be executed when the operation finishes successfully
+     * @param onError Function to be executed when the operation finishes unsuccessfully
+     */
+    getRecommendationsV2(options: RecommendationOptions, onSuccess: (recommendationResponse: RecommendationResponse) => void, onError: (error: Error) => void): void;
     /**
      * This method gets customer's highest-priority screen view campaign.
      *
      * @param onSuccess Function to be executed when the operation finishes successfully
      * @param onError Function to be executed when the operation finishes unsuccessfully
      *
+     * @deprecated Use the new `Content.generateScreenView(feedSlug:onSuccess:onError)` method instead.
      */
     getScreenView(onSuccess: (screenViewResponse: ScreenViewResponse) => void, onError: (error: Error) => void): void;
+    /**
+     * This method generates customer's highest-priority screen view campaign.
+     *
+     * @param onSuccess Function to be executed when the operation finishes successfully
+     * @param onError Function to be executed when the operation finishes unsuccessfully
+     *
+     */
+    generateScreenView(feedSlug: string, onSuccess: (screenView: ScreenView) => void, onError: (error: Error) => void): void;
 }
 export { ContentModule };